+++ config.json
@@ -6,5 +6,5 @@
   "log_format": "%(asctime)s [%(levelname)s] %(name)s - %(message)s",
   "train_percent": 0.7,
   "test_percent": 0.2,
-  "model_path": "./model_2024-03-13_epoch_2.ptn"
+  "model_path": "./data/model_2024-03-13_epoch_2.ptn"
 }
